create app header component
