+++ .vite/deps/@mui_material_FormControl.js
@@ -18,8 +18,7 @@ import {
   require_prop_types,
   styled_default,
   useThemeProps
-} from "./chunk-ALPUVBS4.js";
-import "./chunk-JKGW3AQO.js";
+} from "./chunk-6BF7EGOW.js";
 import {
   __toESM,
   require_react
