+++ .vite/deps/@mui_material_MenuItem.js
@@ -4,7 +4,7 @@ import {
   TransitionGroup_default,
   useEnhancedEffect_default,
   useForkRef_default
-} from "./chunk-BKPP3D3N.js";
+} from "./chunk-53IENX73.js";
 import "./chunk-MDIHFEFG.js";
 import {
   _extends,
@@ -26,8 +26,7 @@ import {
   useEventCallback_default,
   useIsFocusVisible,
   useThemeProps
-} from "./chunk-ALPUVBS4.js";
-import "./chunk-JKGW3AQO.js";
+} from "./chunk-6BF7EGOW.js";
 import {
   __toESM,
   require_react
