+++ .vite/deps/@mui_material_Select.js
@@ -4,7 +4,7 @@ import {
   Transition_default,
   useEnhancedEffect_default as useEnhancedEffect_default2,
   useForkRef_default
-} from "./chunk-BKPP3D3N.js";
+} from "./chunk-53IENX73.js";
 import {
   require_react_dom
 } from "./chunk-MDIHFEFG.js";
@@ -54,8 +54,7 @@ import {
   useThemeProps,
   useTheme_default,
   visuallyHidden_default
-} from "./chunk-ALPUVBS4.js";
-import "./chunk-JKGW3AQO.js";
+} from "./chunk-6BF7EGOW.js";
 import {
   __commonJS,
   __toESM,
