+++ src/styles/globalStyles.ts
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const AppContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+  min-width: 100vw;
+  background-color: #f3f3ff;
+`;
